RUB-309 split safe rust test modules
