+++ clients/rust/crates/rubin-consensus/src/worker_pool.rs
@@ -259,360 +259,5 @@ fn worker_result_missing_failure<R, E>() -> WorkerResult<R, E> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        collect_values, first_error, run_worker_pool, WorkerCancellationToken, WorkerPool,
-        WorkerPoolError, WorkerPoolRunError, FIXED_WORKER_PANIC_MESSAGE,
-    };
-    use std::panic;
-    use std::sync::atomic::{AtomicUsize, Ordering};
-    use std::sync::Arc;
-    use std::time::Duration;
-
-    fn int_identity(_: &WorkerCancellationToken, value: usize) -> Result<usize, String> {
-        Ok(value)
-    }
-
-    fn int_double(_: &WorkerCancellationToken, value: usize) -> Result<usize, String> {
-        Ok(value * 2)
-    }
-
-    fn fail_odd(_: &WorkerCancellationToken, value: usize) -> Result<usize, String> {
-        if value % 2 == 1 {
-            return Err(format!("odd: {value}"));
-        }
-        Ok(value)
-    }
-
-    #[test]
-    fn worker_pool_empty() {
-        let token = WorkerCancellationToken::new();
-        let pool = WorkerPool::new(4, 8, int_identity);
-        let results: Vec<_> = pool.run::<_, _, String>(&token, Vec::new()).unwrap();
-        assert!(results.is_empty());
-    }
-
-    #[test]
-    fn worker_pool_single_task() {
-        let token = WorkerCancellationToken::new();
-        let pool = WorkerPool::new(4, 8, int_double);
-        let results = pool.run(&token, vec![21usize]).unwrap();
-        assert_eq!(results.len(), 1);
-        assert_eq!(results[0].value, Some(42));
-        assert_eq!(results[0].error, None);
-    }
-
-    #[test]
-    fn worker_pool_multiple_tasks_preserve_order() {
-        let token = WorkerCancellationToken::new();
-        let pool = WorkerPool::new(2, 8, int_double);
-        let tasks = vec![1usize, 2, 3, 4, 5];
-        let results = pool.run(&token, tasks.clone()).unwrap();
-        assert_eq!(results.len(), tasks.len());
-        for (idx, result) in results.iter().enumerate() {
-            assert_eq!(result.error, None, "task {idx} unexpectedly failed");
-            assert_eq!(result.value, Some(tasks[idx] * 2));
-        }
-    }
-
-    #[test]
-    fn worker_pool_task_errors_preserve_index_order() {
-        let token = WorkerCancellationToken::new();
-        let pool = WorkerPool::new(4, 8, fail_odd);
-        let tasks = vec![2usize, 3, 4, 5, 6];
-        let results = pool.run(&token, tasks).unwrap();
-
-        assert_eq!(results[0].value, Some(2));
-        assert_eq!(results[0].error, None);
-        assert_eq!(
-            results[1].error,
-            Some(WorkerPoolError::Task("odd: 3".to_string()))
-        );
-        assert_eq!(results[2].value, Some(4));
-        assert_eq!(results[2].error, None);
-        assert_eq!(
-            results[3].error,
-            Some(WorkerPoolError::Task("odd: 5".to_string()))
-        );
-        assert_eq!(results[4].value, Some(6));
-        assert_eq!(results[4].error, None);
-    }
-
-    #[test]
-    fn worker_pool_zero_workers_defaults_to_parallelism() {
-        let token = WorkerCancellationToken::new();
-        let pool = WorkerPool::new(0, 16, int_identity);
-        let tasks = (0usize..10).collect::<Vec<_>>();
-        let results = pool.run(&token, tasks).unwrap();
-        for (idx, result) in results.iter().enumerate() {
-            assert_eq!(result.error, None);
-            assert_eq!(result.value, Some(idx));
-        }
-    }
-
-    #[test]
-    fn worker_pool_workers_capped_at_task_count() {
-        let token = WorkerCancellationToken::new();
-        let current = Arc::new(AtomicUsize::new(0));
-        let max_seen = Arc::new(AtomicUsize::new(0));
-        let pool = WorkerPool::new(100, 8, {
-            let current = Arc::clone(&current);
-            let max_seen = Arc::clone(&max_seen);
-            move |_: &WorkerCancellationToken, value: usize| -> Result<usize, String> {
-                let now = current.fetch_add(1, Ordering::SeqCst) + 1;
-                max_seen.fetch_max(now, Ordering::SeqCst);
-                std::thread::sleep(Duration::from_millis(10));
-                current.fetch_sub(1, Ordering::SeqCst);
-                Ok(value)
-            }
-        });
-
-        let _ = pool.run(&token, vec![1usize, 2, 3]).unwrap();
-        assert!(max_seen.load(Ordering::SeqCst) <= 3);
-    }
-
-    #[test]
-    fn worker_pool_panic_recovery_keeps_other_tasks_running() {
-        let token = WorkerCancellationToken::new();
-        let pool = WorkerPool::new(2, 8, |_: &WorkerCancellationToken, value: usize| {
-            if value == 3 {
-                panic!("task 3 panicked");
-            }
-            Ok::<usize, String>(value * 10)
-        });
-        let tasks = vec![1usize, 2, 3, 4, 5];
-        let results = pool.run(&token, tasks.clone()).unwrap();
-
-        assert_eq!(
-            results[2].error,
-            Some(WorkerPoolError::Panic(
-                FIXED_WORKER_PANIC_MESSAGE.to_string()
-            ))
-        );
-        for (idx, result) in results.iter().enumerate() {
-            if idx == 2 {
-                continue;
-            }
-            assert_eq!(result.error, None);
-            assert_eq!(result.value, Some(tasks[idx] * 10));
-        }
-    }
-
-    #[test]
-    fn worker_pool_context_cancellation_marks_unstarted_tasks() {
-        let token = WorkerCancellationToken::new();
-        let pool = WorkerPool::new(1, 8, {
-            let token = token.clone();
-            move |task_token: &WorkerCancellationToken, value: usize| -> Result<usize, String> {
-                if value == 2 {
-                    token.cancel();
-                    return Ok(value);
-                }
-                if task_token.is_cancelled() {
-                    return Err("cancelled".to_string());
-                }
-                Ok(value * 10)
-            }
-        });
-
-        let results = pool.run(&token, vec![1usize, 2, 3, 4, 5]).unwrap();
-        assert_eq!(results[0].value, Some(10));
-        assert_eq!(results[0].error, None);
-        assert_eq!(results[1].value, Some(2));
-        assert_eq!(results[1].error, None);
-        for result in results.iter().skip(2) {
-            assert_eq!(result.value, None);
-            assert_eq!(result.error, Some(WorkerPoolError::Cancelled));
-        }
-    }
-
-    #[test]
-    fn worker_pool_already_cancelled_token_short_circuits_all_tasks() {
-        let token = WorkerCancellationToken::new();
-        token.cancel();
-        let pool = WorkerPool::new(4, 8, int_identity);
-        let results = pool.run(&token, vec![1usize, 2, 3]).unwrap();
-        for result in results {
-            assert_eq!(result.value, None);
-            assert_eq!(result.error, Some(WorkerPoolError::Cancelled));
-        }
-    }
-
-    #[test]
-    fn worker_pool_run_func_helper() {
-        let token = WorkerCancellationToken::new();
-        let results = run_worker_pool(&token, 2, 8, vec![10usize, 20, 30], int_double).unwrap();
-        let values = collect_values(results).expect("all tasks should succeed");
-        assert_eq!(values, vec![20usize, 40, 60]);
-    }
-
-    #[test]
-    fn worker_pool_first_error_returns_lowest_index_error() {
-        let mut results = vec![
-            super::WorkerResult::success(1usize),
-            super::WorkerResult::success(2usize),
-            super::WorkerResult::success(3usize),
-        ];
-        assert_eq!(first_error(&results), None);
-
-        results[2] = super::WorkerResult::failure(WorkerPoolError::Task("fail at 2".to_string()));
-        assert_eq!(
-            first_error(&results),
-            Some(&WorkerPoolError::Task("fail at 2".to_string()))
-        );
-
-        results[0] = super::WorkerResult::failure(WorkerPoolError::Task("fail at 0".to_string()));
-        assert_eq!(
-            first_error(&results),
-            Some(&WorkerPoolError::Task("fail at 0".to_string()))
-        );
-    }
-
-    #[test]
-    fn worker_pool_collect_values_returns_first_error() {
-        let ok = vec![
-            super::WorkerResult::<usize, String>::success(10usize),
-            super::WorkerResult::<usize, String>::success(20usize),
-            super::WorkerResult::<usize, String>::success(30usize),
-        ];
-        assert_eq!(collect_values(ok).unwrap(), vec![10usize, 20, 30]);
-
-        let bad = vec![
-            super::WorkerResult::<usize, String>::success(10usize),
-            super::WorkerResult::<usize, String>::failure(WorkerPoolError::Task("bad".to_string())),
-            super::WorkerResult::<usize, String>::success(30usize),
-        ];
-        assert_eq!(
-            collect_values(bad),
-            Err(WorkerPoolError::Task("bad".to_string()))
-        );
-    }
-
-    #[test]
-    fn worker_pool_result_order_is_deterministic_across_runs() {
-        for _ in 0..10 {
-            let token = WorkerCancellationToken::new();
-            let pool = WorkerPool::new(4, 32, |_: &WorkerCancellationToken, value: usize| {
-                if value.is_multiple_of(3) {
-                    std::thread::sleep(Duration::from_millis(1));
-                }
-                Ok::<usize, String>(value * value)
-            });
-            let tasks = (0usize..20).collect::<Vec<_>>();
-            let results = pool.run(&token, tasks.clone()).unwrap();
-            for (idx, result) in results.iter().enumerate() {
-                assert_eq!(result.error, None);
-                assert_eq!(result.value, Some(tasks[idx] * tasks[idx]));
-            }
-        }
-    }
-
-    #[test]
-    fn worker_pool_bounded_concurrency() {
-        const MAX_WORKERS: usize = 3;
-        let token = WorkerCancellationToken::new();
-        let current = Arc::new(AtomicUsize::new(0));
-        let max_seen = Arc::new(AtomicUsize::new(0));
-        let pool = WorkerPool::new(MAX_WORKERS, 32, {
-            let current = Arc::clone(&current);
-            let max_seen = Arc::clone(&max_seen);
-            move |_: &WorkerCancellationToken, value: usize| -> Result<usize, String> {
-                let now = current.fetch_add(1, Ordering::SeqCst) + 1;
-                max_seen.fetch_max(now, Ordering::SeqCst);
-                std::thread::sleep(Duration::from_millis(20));
-                current.fetch_sub(1, Ordering::SeqCst);
-                Ok(value)
-            }
-        });
-
-        let tasks = (0usize..20).collect::<Vec<_>>();
-        let _ = pool.run(&token, tasks).unwrap();
-        assert!(max_seen.load(Ordering::SeqCst) <= MAX_WORKERS);
-    }
-
-    #[test]
-    fn worker_pool_large_batch() {
-        let token = WorkerCancellationToken::new();
-        let tasks = (0usize..1000).collect::<Vec<_>>();
-        let pool = WorkerPool::new(8, 1000, int_double);
-        let results = pool.run(&token, tasks.clone()).unwrap();
-        assert_eq!(results.len(), tasks.len());
-        for (idx, result) in results.iter().enumerate() {
-            assert_eq!(result.error, None);
-            assert_eq!(result.value, Some(tasks[idx] * 2));
-        }
-    }
-
-    #[test]
-    fn worker_pool_rejects_zero_max_tasks() {
-        let token = WorkerCancellationToken::new();
-        let pool = WorkerPool::new(4, 0, int_identity);
-        assert_eq!(
-            pool.run(&token, vec![1usize, 2, 3]),
-            Err(WorkerPoolRunError::InvalidMaxTasks)
-        );
-    }
-
-    #[test]
-    fn worker_pool_rejects_task_count_over_limit() {
-        let token = WorkerCancellationToken::new();
-        let pool = WorkerPool::new(4, 2, int_identity);
-        assert_eq!(
-            pool.run(&token, vec![1usize, 2, 3]),
-            Err(WorkerPoolRunError::TooManyTasks {
-                task_count: 3,
-                max_tasks: 2,
-            })
-        );
-    }
-
-    #[test]
-    fn worker_pool_non_string_panic_uses_fixed_message() {
-        let token = WorkerCancellationToken::new();
-        let pool = WorkerPool::new(1, 1, |_: &WorkerCancellationToken, _value: usize| {
-            panic::panic_any(7usize);
-            #[allow(unreachable_code)]
-            Ok::<usize, String>(0)
-        });
-
-        let results = pool.run(&token, vec![1usize]).unwrap();
-        assert_eq!(
-            results[0].error,
-            Some(WorkerPoolError::Panic(
-                FIXED_WORKER_PANIC_MESSAGE.to_string()
-            ))
-        );
-    }
-
-    #[test]
-    fn worker_pool_error_display_variants() {
-        assert_eq!(
-            WorkerPoolError::<String>::Cancelled.to_string(),
-            "context canceled"
-        );
-        assert_eq!(
-            WorkerPoolError::<String>::Panic(FIXED_WORKER_PANIC_MESSAGE.to_string()).to_string(),
-            "worker panic"
-        );
-        assert_eq!(
-            WorkerPoolError::<String>::Task("task failed".to_string()).to_string(),
-            "task failed"
-        );
-    }
-
-    #[test]
-    fn worker_pool_run_error_display_variants() {
-        assert_eq!(
-            WorkerPoolRunError::InvalidMaxTasks.to_string(),
-            "worker pool max_tasks must be positive"
-        );
-        assert_eq!(
-            WorkerPoolRunError::TooManyTasks {
-                task_count: 7,
-                max_tasks: 3,
-            }
-            .to_string(),
-            "worker pool task count 7 exceeds configured limit 3"
-        );
-    }
-}
+#[path = "tests/worker_pool.rs"]
+mod tests;
